+++ config/env.docker.sh
@@ -27,3 +27,5 @@ export HELP_LINK='disable'
 export ASSETS_URL='https://url.com'
 # Matomo tracking URL (use disable to desactivate tracking)
 export MATOMO_TRACKER_URL='disable'
+export MAX_FAILED_LOGIN_ATTEMPTS='5'
+export DELAY_BEFORE_NEXT_LOGIN_ATTEMPT='1'